Rewrite thumbnail-proxy using got
Add Cache-Control headers (30 days) to keep the thumbnails in the
browser cache.
