+++ lib/api/controllers/records.js
@@ -1,7 +1,9 @@
 const mongoose = require('mongoose')
 const {clone} = require('lodash')
+
 const search = require('../../search')
-const {proxyThumbnail} = require('../../utils/thumbnail-proxy')
+
+const proxyThumbnail = require('../middlewares/thumbnail-proxy')
 
 const ConsolidatedRecord = mongoose.model('ConsolidatedRecord')
 const RecordRevision = mongoose.model('RecordRevision')
@@ -81,8 +83,10 @@ exports.consolidate = function (req, res, next) {
 
 exports.thumbnail = function (req, res) {
   const thumbnail = req.record.metadata.thumbnails.find(th => th.originalUrlHash === req.params.originalUrlHash)
+
   if (!thumbnail) {
     return res.sendStatus(404)
   }
-  proxyThumbnail(thumbnail.originalUrl, req, res)
+
+  proxyThumbnail(thumbnail.originalUrl)(req, res)
 }
