+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@etalab/fr-bounding-boxes": "^0.1.3",
     "@turf/turf": "^5.1.6",
+    "@tusbar/cache-control": "^0.2.1",
     "bbox-match": "^0.1.0",
     "bluebird": "^3.3.0",
     "connect-mongo": "^2.0.0",
